% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheatcode.R
\name{cheatcode}
\alias{cheatcode}
\title{Helps with cheating by producing helpful information}
\usage{
cheatcode(help_with = c("ggplot", "rnorm", "apply"))
}
\arguments{
\item{help_with}{}
}
\value{
a string containing helpful information
}
\description{
Helps with cheating by producing helpful information
}
