% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbowGraph.R
\name{rainbowGraph}
\alias{rainbowGraph}
\title{Create a rainbow coloured line graph}
\usage{
rainbowGraph(dataframe, x, y, title = "My Beautiful Plot")
}
\arguments{
\item{dataframe}{a dataframe containing your data, passed to \code{\link{ggplot}}}

\item{x}{one variable of interest}

\item{y}{another variable of interest}

\item{title}{title that should be given to the graph}
}
\value{
a rainbow line graph
}
\description{
Create a rainbow coloured line graph
}
